requirements.txt
Mastodon.py
tweepy
pyyaml
